requirements.txt
pandas>=1.3.0
nltk>=3.6.0
PyYAML>=5.4.0
requests>=2.26.0
PyGithub>=1.55
tqdm>=4.62.0